+++ docs/source/conf.py
@@ -51,6 +51,7 @@ autosummary_ignore_module_all = False
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
 napoleon_attr_annotations = True
+napoleon_use_rtype = False
 
 # todo config
 todo_include_todos = True
@@ -97,7 +98,7 @@ myst_url_schemes = {
 
 # -- HTML output ------------------------------------------------------------
 html_theme = "sphinx_immaterial"
-html_title = f"{project} v{release}"
+html_title = project
 html_theme_options = {
     "repo_url": "https://github.com/cherab/imas",
     "repo_name": "CHERAB-IMAS",